# doc/api.yml
openapi: 3.0.0
info:
  description: "Tasktracker API Documentation"
  version: "1.0.0"
  title: Tasktracker
  contact:
    email: kot8988@yandex.ru
servers: []
tags:
  - name: user
    description: Operations about user
  - name: task
    description: Operations with tasks
  - name: events
    description: Operations with tasks events (views and likes)
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user by login and password
      operationId: CreateUser
      requestBody:
        required: true
        description: A JSON object containing the login and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: User was successfully created
        '400':
          description: Bad request
        '409':
          description: User already exists
        '5XX':
          description: Internal server error
  /user/login:
    get:
      tags:
        - user
      summary: Log user into the system and returns auth cookie
      operationId: LoginUser
      parameters:
        - name: login
          in: query
          description: user login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: user password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >
            Successfully authenticated.
            The session ID is returned in a cookie named `JSESSIONID`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
        '5XX':
          description: Internal server error
  /user/{login}:
    put:
      tags:
        - user
      summary: Update user profile
      description: Update existing user information ignoring empty fields
      operationId: UpdateUser
      parameters:
        - name: login
          in: path
          description: user login
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Updated user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
      responses:
        '200':
          description: User info successfully updated
        '400':
          description: Invalid user data supplied
        '401':
          description: Access token is missing or invalid
        '403':
          description: Trying to update another user data
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
  /task:
    post:
      tags:
        - task
      summary: Create task
      description: Create task by description and status
      operationId: CreateTask
      requestBody:
        required: true
        description: A JSON object containing task description and status.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreds'
        '400':
          description: Invalid task data supplied
        '401':
          description: Access token is missing or invalid
        '403':
          description: Trying to create task under other username
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
  /task/{id}:
    put:
      tags:
        - task
      summary: Update task
      description: Update task description or/and status by id, empty fields won't be updated
      operationId: UpdateTask
      parameters:
        - name: id
          in: path
          description: task id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        description: A JSON object containing task description and status.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        '400':
          description: Invalid task data supplied
        '401':
          description: Access token is missing or invalid
        '403':
          description: Trying to access other user task
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
    delete:
      tags:
        - task
      summary: Delete task
      description: Delete task by id
      operationId: DeleteTask
      parameters:
        - name: id
          in: path
          description: task id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        '400':
          description: Task with such id not exists
        '401':
          description: Access token is missing or invalid
        '403':
          description: Trying to access other user task
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
    get:
      tags:
        - task
      summary: Get task
      description: Get task by id
      operationId: GetTask
      parameters:
        - name: id
          in: path
          description: task id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfuly get task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Task with such id not exists
        '401':
          description: Access token is missing or invalid
        '403':
          description: Trying to access other user task
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
  /task/page:
    get:
      tags:
        - task
      summary: Get tasks with pagination
      description: Get tasks by page number and number of tasks in page
      operationId: GetTasksByPage
      parameters:
        - name: user
          in: query
          description: owner of posts to get
          required: true
          schema:
            type: string
        - name: size
          in: query
          description: number of posts in page
          required: true
          schema:
            type: integer
        - name: number
          in: query
          description: page number
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfuly get list of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: Invalid PageRequest
        '401':
          description: Access token is missing or invalid
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
  /view:
    post:
      tags:
        - events
      summary: Add user view to task
      description: Add one user view to task by task id and user id
      operationId: ViewTask
      parameters:
        - name: task_id
          in: query
          description: task id to view
          schema:
            type: integer
      responses:
        '200':
          description: Succesfully add user view to task
        '400':
          description: Invalid task id
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []
  /like:
    post:
      tags:
        - events
      summary: Add user like to task
      description: Add one user like to task by task id and user id
      operationId: LikeTask
      parameters:
        - name: task_id
          in: query
          description: task id to like
          schema:
            type: integer
      responses:
        '200':
          description: Succesfully add user like to task
        '400':
          description: Invalid task id
        '5XX':
          description: Internal server error
      security:
        - cookieAuth: []

components:
  schemas:
    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    UserData:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
          format: email
        phone:
          type: string
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        owner:
          type: string
        body:
          type: string
        status:
          type: integer
          enum: [0, 1, 2, 3]
          x-enum-varnames: [DEFAULT, OPEN, IN_PROGRESS, COMPLETED]
    TaskCreds:
      type: object
      properties:
        id:
          type: integer
          format: int64
        owner:
          type: string
    Empty:
      type: object
      properties:
        message:
          type: string
    PageResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID